add RepairServiceApp
